fix: 补充GroupInfo的修正代码
