+++ migrate/v2/v150/v150_attrs.go
@@ -111,9 +111,9 @@ func attrsGroupUserMigrate(db *gorm.DB) (int, int, error) {
 			countFailed++
 			continue
 		}
-		// 判断data是否为正常JSON字符串，以及ID是否存在，若不行的话也需要跳过
-		if !gjson.Valid(string(row.Data)) || row.ID == "" {
-			log.Warnf("[损坏数据] 跳过一行数据，用户核心数据已经损坏: %v", err)
+		// 跳过：① 无效 JSON ② 不是 JSON 对象 ③ ID 为空
+		if res := gjson.ParseBytes(row.Data); !res.IsObject() || row.ID == "" {
+			log.Warnf("[损坏数据] 跳过一行数据 (%v)-(%s)，用户GroupUser核心数据已经损坏", row.ID, string(row.Data))
 			countFailed++
 			continue
 		}
@@ -213,9 +213,9 @@ func attrsGroupMigrate(db *gorm.DB) (int, int, error) {
 			countFailed++
 			continue
 		}
-		// 判断data是否为正常JSON字符串，以及ID是否存在，若不行的话也需要跳过
-		if !gjson.Valid(string(row.Data)) || row.ID == "" {
-			log.Warnf("[损坏数据] 跳过一行数据，用户核心数据已经损坏: %v", err)
+		// 跳过：① 无效 JSON ② 不是 JSON 对象 ③ ID 为空
+		if res := gjson.ParseBytes(row.Data); !res.IsObject() || row.ID == "" {
+			log.Warnf("[损坏数据] 跳过一行数据 (%v)-(%s)，用户Group核心数据已经损坏", row.ID, string(row.Data))
 			countFailed++
 			continue
 		}
@@ -289,9 +289,9 @@ func attrsUserMigrate(db *gorm.DB) (int, int, int, error) {
 			countFailed++
 			continue
 		}
-		// 判断data是否为正常JSON字符串，以及ID是否存在，若不行的话也需要跳过
-		if !gjson.Valid(string(row.Data)) || row.ID == "" {
-			log.Warnf("[损坏数据] 跳过一行数据，用户核心数据已经损坏: %v", err)
+		// 跳过：① 无效 JSON ② 不是 JSON 对象 ③ ID 为空
+		if res := gjson.ParseBytes(row.Data); !res.IsObject() || row.ID == "" {
+			log.Warnf("[损坏数据] 跳过一行数据 (%v)-(%s)，用户核心数据已经损坏", row.ID, string(row.Data))
 			countFailed++
 			continue
 		}
@@ -538,21 +538,21 @@ func dataDBInit(dboperator operator.DatabaseOperator, logf func(string)) error {
 func createMySQLIndexForLogInfo(db *gorm.DB) (err error) {
 	// 创建前缀索引
 	// 检查并创建索引
-	if !db.Migrator().HasIndex(&LogInfoHookMySQL{}, "idx_log_name") {
+	if !db.Migrator().HasIndex(&model.LogInfoHookMySQL{}, "idx_log_name") {
 		err = db.Exec("CREATE INDEX idx_log_name ON logs (name(20));").Error
 		if err != nil {
 			log.Errorf("创建idx_log_name索引失败,原因为 %v", err)
 		}
 	}
 
-	if !db.Migrator().HasIndex(&LogInfoHookMySQL{}, "idx_logs_group") {
+	if !db.Migrator().HasIndex(&model.LogInfoHookMySQL{}, "idx_logs_group") {
 		err = db.Exec("CREATE INDEX idx_logs_group ON logs (group_id(20));").Error
 		if err != nil {
 			log.Errorf("创建idx_logs_group索引失败,原因为 %v", err)
 		}
 	}
 
-	if !db.Migrator().HasIndex(&LogInfoHookMySQL{}, "idx_logs_updated_at") {
+	if !db.Migrator().HasIndex(&model.LogInfoHookMySQL{}, "idx_logs_updated_at") {
 		err = db.Exec("CREATE INDEX idx_logs_updated_at ON logs (updated_at);").Error
 		if err != nil {
 			log.Errorf("创建idx_logs_updated_at索引失败,原因为 %v", err)
@@ -564,13 +564,13 @@ func createMySQLIndexForLogInfo(db *gorm.DB) (err error) {
 func createMySQLIndexForLogOneItem(db *gorm.DB) (err error) {
 	// 创建前缀索引
 	// 检查并创建索引
-	if !db.Migrator().HasIndex(&LogOneItemHookMySQL{}, "idx_log_items_group_id") {
+	if !db.Migrator().HasIndex(&model.LogOneItemHookMySQL{}, "idx_log_items_group_id") {
 		err = db.Exec("CREATE INDEX idx_log_items_group_id ON log_items(group_id(20))").Error
 		if err != nil {
 			log.Errorf("创建idx_logs_group索引失败,原因为 %v", err)
 		}
 	}
-	if !db.Migrator().HasIndex(&LogOneItemHookMySQL{}, "idx_raw_msg_id") {
+	if !db.Migrator().HasIndex(&model.LogOneItemHookMySQL{}, "idx_raw_msg_id") {
 		err = db.Exec("CREATE INDEX idx_raw_msg_id ON log_items(raw_msg_id(20))").Error
 		if err != nil {
 			log.Errorf("创建idx_log_group_id_name索引失败,原因为 %v", err)
@@ -578,46 +578,6 @@ func createMySQLIndexForLogOneItem(db *gorm.DB) (err error) {
 	}
 	// MYSQL似乎不能创建前缀联合索引，放弃所有的前缀联合索引
 	return nil
-}
-
-type LogInfoHookMySQL struct {
-	ID         uint64  `json:"id" gorm:"primaryKey;autoIncrement;column:id"`
-	Name       string  `json:"name" gorm:"column:name"`
-	GroupID    string  `json:"groupId" gorm:"column:group_id"`
-	CreatedAt  int64   `json:"createdAt" gorm:"column:created_at"`
-	UpdatedAt  int64   `json:"updatedAt" gorm:"column:updated_at"`
-	Size       *int    `json:"size" gorm:"<-:false"`
-	Extra      *string `json:"-" gorm:"column:extra"`
-	UploadURL  string  `json:"-" gorm:"column:upload_url"`
-	UploadTime int     `json:"-" gorm:"column:upload_time"`
-}
-
-func (*LogInfoHookMySQL) TableName() string {
-	return "logs"
-}
-
-type LogOneItemHookMySQL struct {
-	ID             uint64      `json:"id" gorm:"primaryKey;autoIncrement;column:id"`
-	LogID          uint64      `json:"-" gorm:"column:log_id"`
-	GroupID        string      `gorm:"column:group_id"`
-	Nickname       string      `json:"nickname" gorm:"column:nickname"`
-	IMUserID       string      `json:"IMUserId" gorm:"column:im_userid"`
-	Time           int64       `json:"time" gorm:"column:time"`
-	Message        string      `json:"message"  gorm:"column:message"`
-	IsDice         bool        `json:"isDice"  gorm:"column:is_dice"`
-	CommandID      int64       `json:"commandId"  gorm:"column:command_id"`
-	CommandInfo    interface{} `json:"commandInfo" gorm:"-"`
-	CommandInfoStr string      `json:"-" gorm:"column:command_info"`
-	RawMsgID       interface{} `json:"rawMsgId" gorm:"-"`
-	RawMsgIDStr    string      `json:"-" gorm:"column:raw_msg_id"`
-	UniformID      string      `json:"uniformId" gorm:"column:user_uniform_id"`
-	Channel        string      `json:"channel" gorm:"-"`
-	Removed        *int        `gorm:"column:removed" json:"-"`
-	ParentID       *int        `gorm:"column:parent_id" json:"-"`
-}
-
-func (*LogOneItemHookMySQL) TableName() string {
-	return "log_items"
 }
 
 func logDBInit(dboperator operator.DatabaseOperator, logf func(string)) error {
